Update patch set 2
Patch Set 2: Code-Review+2

Actually, the Qcom documents reveal partial update not being as awesome as it sounds (it *does* bring power usage improvements, but not that much with a typical usecase).
Let's continue debugging this but ship the workaround if needed.
Since this apparently works for Google/non-BSP cases, I assume the kernel space is not broken, thus we disable it here and not there.

Patch-set: 2
Label: Code-Review=+2