sdm: core: Update the mixer, framebuffer and display properly
This is needed for the HAL to properly send the mixer and
display properties correctly to the kernel, for it to use
those informations to calculate the right pixel clock for
the new resolution.

If we don't send those informations, the kernel driver
will calculate the pixclock for the old resolution,
resulting in insufficient bandwidth, loss of commands, bad
timings and finally hardware faults because of bad and/or
unexpected DSI status, which will then obviously lead to
device reboot.

Change-Id: Ied222fdd07bcbac00df106397828c2a2a2aa170a