Update patch set 1
Patch Set 1: Code-Review+1

This is fine with me. The default configuration to have MPD render this instead isn't optimal as it uses up CPU cycles that should be sent the GPUs way. This partially mimics forcing 2d rendering on the GPU, but only with critical values. I know this is a bug as enabling the setting appears to increase overall UX fluidity, and this accomplishes the same

Patch-set: 1
Reviewer: Gerrit User 1000166 <1000166@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Code-Review=+1