sdm: drm: Set CRTC mode, only if display config changes
When display mode is set on a CRTC object, it is cached by DRM driver
and it is not required to be programmed in Validate or Commit of each
draw cycle, until the display configuration gets updated. Hence, set
the CRTC mode, only if the display configuration or state changes.

CRs-Fixed: 2240899
Change-Id: I89cef5c67c210fc070d0aa74d460e362a08a78fe