lights: supress errors entering VR mode on sdm845
VTS requires that the lights HAL only report errors if the device
does not support VR mode at all.  On sdm845, the kernel can also
report an error if the display is not in the proper power state
for VR entry.

To allow VTS to pass, this patch suppresses these new errors.
However, it continues to warn in the log.  If the VR device is not
present, the HAL contniues to produce the expected error.

Bug: 80028985
Test: VtsHalLightV2_0TargetTest passes on Blueline and Crosshatch
Change-Id: Iaa2d598b16c5f56175e3406a588e504722c0b657