Update patch set 1
Patch Set 1:

As with LineageOS, we may want this patch picked on other SoC's branches as YCrCb could be used by other OEMs. AFAIAA the LG V50 ThinQ also outputs it's camera preview stream in YCrCb, as opposed to the "usual" format of YCbCr. The LG V50 ThinQ runs on the Snapdragon 855 (SM8150) platform and so would need this patch in the quartz-855 branch too.

The follow-up patch for LG's VENUS format could be avoided where it isn't necessary, but there's no harm in adding it in.

Patch-set: 1